+++ client/front/lib/errand_check.dart
@@ -6,8 +6,8 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_naver_map/flutter_naver_map.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:front/show_errand.dart';
 import 'package:intl/intl.dart';
+import 'mainshowerrand/mainshowerrand.dart';
 import 'request.dart';
 import 'main_post_page.dart';
 import 'package:http/http.dart' as http;
